+++ compiler-core/src/type_/tests/warnings.rs
@@ -5255,3 +5255,17 @@ pub fn thingy(wibble) {
     "#
     );
 }
+
+#[test]
+fn empty_string_in_bit_array_pattern_segment() {
+    assert_warning!(
+        r#"
+pub fn go(x) {
+  case x {
+    <<"">> -> True
+    _ -> False
+  }
+}
+"#
+    );
+}