+++ CHANGELOG.md
@@ -153,6 +153,10 @@
   through further usages.
   ([James Dolan](https://github.com/jamesdolan16))
 
+- Matching on bit array patterns which contain empty string segments is now
+  deprecated.
+  ([Surya Rose](https://github.com/GearsDatapacks))
+
 ### Build tool
 
 - The build tool now stores its build cache in a more compact binary format,